requirements.txt
Flask
elastic-apm[flask]
python-dotenv
elastic-opentelemetry